smart-ingest-kit/requirements.txt
docling
pydantic
loguru
llama-index-core  # Optional, but good for typing
